Ensure teardown is called to drop static wrapped areana in example component test
